pyproject.toml
[project]
name = "satin"
version = "0.1.0"
description = "A Simple Annotation Tool for Images."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[granian]>=0.116.1",
    "granian>=2.4.2",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.13.2",
    "strawberry-graphql[fastapi]>=0.278.0",
]

[project.scripts]
satin = "satin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "granian[reload]>=2.4.2",
    "httpx>=0.28.1",
    "hypothesis>=6.136.4",
    "isort>=6.0.1",
    "mongomock-motor>=0.0.36",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-httpx>=0.34.0",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.5",
    "sphinx>=8.2.1",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-autodoc-typehints>=2.5.1",
    "myst-parser>=4.0.0",
    "ssort>=0.15.0",
    "strawberry-graphql[cli]>=0.278.0",
]

###########
# Linting #
###########

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
exclude = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py313"
lint.select = [
  "ALL", # include all the rules, including new ones
]
lint.ignore = [
  "ANN",     # flake8-annotations
  "D100",    # ignore missing docs
  "D101",
  "D203",    # ignore no blank lines before class docstring
  "D213",    # multi-line docstring summary should start at the second line
  "COM812",  # ignore missing blank line after last section
  "FBT001",  # ignore boolean arguments
  "FBT002",
  "TD",      # ignore TODOs
  "FIX002",  # ignore TODOs
  "PLR0913", # ignore too many arguments
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
  "D104", # ignore missing docs
]
"tests/**/*.py" = [
  "D100",    # ignore missing docs
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D107",
  "ANN",     # flake8-annotations
  "S101",    # ignore asserts
  "PLR2004", # ignore magic values
  "PD901",   # allow df as variable name
  "SLF001",  # private method accessed outside of class
  "PGH003",  # allow unspecified type ignores
]
"docs/source/conf.py" = ["ALL"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
